♻️ refactor: Creating components to organize pages forms

+++ StockTracker/Models/Shared/Form/Inputs/NumberInputModel.cs
@@ -2,8 +2,8 @@
 {
   public class NumberInputModel : InputModel
   {
-    public int? Min { get; set; }
-    public int? Max { get; set; }
+    public float? Min { get; set; }
+    public float? Max { get; set; }
     public float Step { get; set; }
 
     public NumberInputModel()
